Get rid of logger spam

+++ NewHorizons/External/OrbitModule.cs
@@ -10,7 +10,7 @@ namespace NewHorizons.External
     public class OrbitModule : Module
     {
         public int SemiMajorAxis { get; set; } = 5000;
-        public int Inclination { get; set; }
+        public float Inclination { get; set; }
         public string PrimaryBody { get; set; }
         public bool IsMoon { get; set; }
         public float LongitudeOfAscendingNode { get; set; }
